+++ app/Http/Requests/Task/CreateTaskRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Project;
+namespace App\Http\Requests\Task;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class CreateProjectRequest extends FormRequest
+class CreateTaskRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,9 +24,8 @@ class CreateProjectRequest extends FormRequest
     public function rules()
     {
         return [
-            'title' => ['required', 'string', 'max:255'],
-            'description' => ['required'],
-            'notes' => ['max:255'],
+            'body' => ['required'],
+            'project_id' => ['numeric', 'exists:projects,id'],
         ];
     }
 }