+++ database/factories/TaskFactory.php
@@ -3,17 +3,18 @@
 namespace Database\Factories;
 
 use App\Models\Project;
+use App\Models\Task;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class ProjectFactory extends Factory
+class TaskFactory extends Factory
 {
     /**
      * The name of the factory's corresponding model.
      *
      * @var string
      */
-    protected $model = Project::class;
+    protected $model = Task::class;
 
     /**
      * Define the model's default state.
@@ -23,9 +24,8 @@ class ProjectFactory extends Factory
     public function definition()
     {
         return [
-            'title' => $this->faker->sentence(),
-            'description' => $this->faker->paragraph(),
-            'notes' => $this->faker->paragraph(),
+            'body' => $this->faker->paragraph(),
+            'project_id' => Project::factory()->create()->id,
             'user_id' => User::factory()->create()->id,
         ];
     }