+++ tests/Feature/Project/CreateProjectTest.php
@@ -11,15 +11,9 @@ class CreateProjectTest extends TestCase
 {
     use RefreshDatabase;
 
-    public function setUp(): void
-    {
-        parent::setUp();
-
-        $this->user = User::factory()->create();
-    }
     public function testCreateProjectView()
     {
-        $this->actingAs($this->user);
+        $this->signIn();
 
         $response = $this->get('projects/create');
 
@@ -30,7 +24,7 @@ class CreateProjectTest extends TestCase
     {
         $this->withoutExceptionHandling();
 
-        $this->actingAs($this->user);
+        $this->signIn();
 
         $project = Project::factory()->make();
 
@@ -40,6 +34,7 @@ class CreateProjectTest extends TestCase
         $this->assertDatabaseHas('projects', [
             'title'=>$project->title,
             'description'=>$project->description,
+            'notes'=>$project->notes,
         ]);
 
         $this->get('/projects')->assertSee($project->title);
@@ -47,7 +42,7 @@ class CreateProjectTest extends TestCase
 
     public function testProjectRequiresTitle()
     {
-        $this->actingAs($this->user);
+        $this->signIn();
 
         $project = Project::factory()->make(['title'=>null]);
 
@@ -57,7 +52,7 @@ class CreateProjectTest extends TestCase
 
     public function testProjectRequiresDescription()
     {
-        $this->actingAs($this->user);
+        $this->signIn();
 
         $project = Project::factory()->make(['description'=>null]);