+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Release Notes
+## v0.1.0 (2021-08-16)
+
+#### New Features:
+
+- [#2](https://github.com/pasanks/managely/issues/2) Create new project
+- [#9](https://github.com/pasanks/managely/issues/9) Add / update project tasks
+- [#12](https://github.com/pasanks/managely/issues/12) Add general notes to the project
+- [#5](https://github.com/pasanks/managely/issues/5) UI to create project / dashboard / single project view
+- [#4](https://github.com/pasanks/managely/issues/4) Add readme file
+- [#6](https://github.com/pasanks/managely/issues/6) Add licence
 
 ## v0.0.1 (2021-08-12)